+++ app/src/main/java/com/spendscan/data/models/MockData.kt
@@ -1,0 +1,76 @@
+package com.spendscan.data.models
+
+val expensesList = listOf(
+    Expenses(
+        id = "1",
+        title = "Аренда квартиры",
+        subtitle = null,
+        createdAt = "2025-06-12T10:00:00Z",
+        trailTag = "",
+        trailText = "100 000 ₽",
+        iconTag = "\uD83C\uDFE0"
+    ),
+    Expenses(
+        id = "2",
+        title = "Одежда",
+        subtitle = null,
+        createdAt = "2025-06-11T12:30:00Z",
+        trailTag = "",
+        trailText = "100 000 ₽",
+        iconTag = "\uD83D\uDC57"
+    ),
+    Expenses(
+        id = "3",
+        title = "На собачку",
+        subtitle = "Джек",
+        createdAt = "2025-06-10T15:45:00Z",
+        trailTag = "",
+        trailText = "100 000 ₽",
+        iconTag = "\uD83D\uDC36"
+    ),
+    Expenses(
+        id = "4",
+        title = "На собачку",
+        subtitle = "Энни",
+        createdAt = "2025-06-09T09:15:00Z",
+        trailTag = "",
+        trailText = "100 000 ₽",
+        iconTag = "\uD83D\uDC36"
+    ),
+    Expenses(
+        id = "5",
+        title = "Ремонт квартиры",
+        subtitle = null,
+        createdAt = "2025-06-08T11:00:00Z",
+        trailTag = "",
+        trailText = "100 000 ₽",
+        iconTag = "РК"
+    ),
+    Expenses(
+        id = "6",
+        title = "Продукты",
+        subtitle = null,
+        createdAt = "2025-06-07T16:20:00Z",
+        trailTag = "",
+        trailText = "100 000 ₽",
+        iconTag = "\uD83C\uDF6D"
+    ),
+    Expenses(
+        id = "7",
+        title = "Спортзал",
+        subtitle = null,
+        createdAt = "2025-06-06T08:40:00Z",
+        trailTag = "",
+        trailText = "100 000 ₽",
+        iconTag = "\uD83C\uDFCB\uFE0F"
+    ),
+    Expenses(
+        id = "8",
+        title = "Медицина",
+        subtitle = null,
+        createdAt = "2025-06-05T14:00:00Z",
+        trailTag = "",
+        trailText = "100 000 ₽",
+        iconTag = "\uD83D\uDC8A"
+    )
+)
