+++ app/src/main/java/com/spendscan/ui/theme/Color.kt
@@ -2,9 +2,10 @@ package com.spendscan.ui.theme
 
 import androidx.compose.ui.graphics.Color
 
-val GreenMain = Color(0xFF29E881)
+val GreenMain = Color(0xFF2AE881)
 val GreenSecond = Color(0xFFd5fae7)
 val Black = Color(0xFF000000)
 val White = Color(0xFFFFFFFF)
 val GreyLight = Color(0xFFCAC4D0)
 val Greyback =Color(0xFFFEF7FF)
+val DarkGrey = Color(0xFF1D1B20)
